Start of RefactoringAssistant. It can now: * move attributes and operations between the classes in a class hierarchy *show (not yet change) opperation properties (atts props should come soon) *create new super/derived classes  (the classes are created and the association between the two classifiers is made, but the association doesnt show in the diagram... I think this has to do with the UMLAssociation Vs. AssociationWidget problem
-> to do:
You can only call the assistant from the context menu of the new (experimental) class widget. to do: add this to normal classwidget context menu and/or tree view's

When refactoring, changes are inmediatly shown in the new classwidget (they get updated
automatically when the UMLClass is modified), but "normal" ClassWidgets dont get updated,
even though the modified signal is emited by the UMLClass. to do: fix this.

to do: a lot of other stuff.. like I said, this is only the start

svn path=/trunk/kdesdk/umbrello/; revision=217458