introduced getter for FontMetrics. This allows to use the correct FontMetrics when printing and thus printed widgets will look like their screen counterpart when the printer font has different metrics than the screen font (wiget size is computed from the font metrics).
Furthermore this change should rather increase performance than decrease it
because QFontMetrics are now cached and only created when something changed.

Fixed layout issues for classwidgets and interfacewidgets where text was
cut off sometimes by using the font metrics for calculations that correspond
to the really used font for rendering.

The font metrics will be updated prior to printing and afterwards.
The remaining problem is that the widgets itself are correct now, but the
spacial relationship between them looks wrong, because those values are
absolute and not computed from the font metrics.

svn path=/trunk/kdesdk/umbrello/; revision=229001