Fix for Bug 54822 by Jean-Louis Houberdon
If you want to understand the problem of the bug to be sure it answers it I can say that texts when loaded have a menu as if they where part of association
(i.e association name, so tr_name) and th constructor where are this lines is
called when floating text is loaded, so by default, they are not the good type.
For text of associations, the same constructor is called, but a setRole is just
called after, so it cannot break association loading, that's why, it should
work perfectly.

svn path=/trunk/kdesdk/umbrello/; revision=231137