Fix for Bug 59190:
Static variables were marked as final in the generated source code. Final and static are not the same. A final variable can only be changed one time in the program, a static variable can be changed as often as you like, but in each object of a class the static variable has allways the same value, because it only exists one time for all instances of a class.

We should think if getter/setter methods for static variable should also be static!

CCMAIL: 59190-close@bugs.kde.org

svn path=/trunk/kdesdk/umbrello/; revision=231721