first try to fix bug 55051: When using the class properties dialog to add a new operation/attribute, the documentation 	entered is now saved before a new operation is created or the properties of the operation/attribute are called.
Well, the fix only works if the selected attribute/operation with the new documentation isn't the last one in the attribute/operation list. If it is the last one, the documentation is still lost when clicking on the Create button. No idea why, it must be some kind of index problem...

svn path=/trunk/kdesdk/umbrello/; revision=242372