fix bug 63248: patch provided by Jean-Louis Houberdon
When you remove the widget (i.e. here the class widget), you remove association widgets too in the same time. So, when you come back to the loop and ask for "obj->getID()" on the next widget, it is possible you have deleted it meanwhile. So, it's necessary to update the object list to be sure you don't have a widget which doesn't exist.

svn path=/trunk/kdesdk/umbrello/; revision=247188