fix bug 57490:
the following 2 items are now fixed in CVS:
> 1) Wrong anchor point deleted
> 3) Anchor point boxes don't go away

I won't fix the following part of the bug report:
> 2) Self-associations-- can delete all interior anchor
> points
> * Make class
> * Make (unidirectional) self-association to that
> class, with rectangular arrow with two non-endpoint anchor points.
> You can delete both of these non-endpoint anchor points, resulting in
> a very odd arrow. Any labels on the arrow are left dangling.

Because if the user really wants to do so, his problem. Even if we restrict the number of points of self-associations to be at least 3, he can always move the middle point on a line between the outer points. This would be the same like deleting the middle point.

svn path=/trunk/kdesdk/umbrello/; revision=249714