Please revert this small change, if you have good arguments why this doesn't improve the handling:
When right clicking on the background of a diagram, the default tool is selected from the toolbar and the context menu is not shown. You have to right click one more time to get to the context menu.
This is only valid, if there is really no widget like a class under the mouse pointer. In this case the right click menu is always shown!

svn path=/trunk/kdesdk/umbrello/; revision=249727