fix for bug #67277 by Daniel Richard G. <skunk@iskunk.org>
Richard:
"The problem is that the FlexLexer.h header has a strange way of handling multiple inclusions of itself. tokenizer.l (tokenizer.cc) and ClassParser.cc both include this header indirectly, and in separate compilation they work fine; but when compiling concatenated source, a "#define yyFlexLexer yyFlexLexer" declaration (inserted by flex) carries through from tokenizer.cc to ClassParser.cc, and causes havoc on the second inclusion. (In the error message quoted above, the two FlexLexer.h lines are the second and first inclusions, respectively.)

The solution is to "#undef yyFlexLexer" at the end of tokenizer."

svn path=/trunk/kdesdk/umbrello/; revision=264733