fix bug #73042: code generation ignores output directory
In fact, it ignored all settings before code generation was started. I also re-added the check, if the output directory exists and is writable. I will backport this change to KDE 3.2 in the next days. Please have a look at the function CodeGenerationWizard::generator()! What was the purpose of this function? Can we remove it completely???

svn path=/trunk/kdesdk/umbrello/; revision=284187