Backport commit from 12 Feb --
CVS commit by thomas:

bug fix: operations NOT being added/removed properly to UMLDoc;
restored the connection between add/removeChildObject and slot
add/removeUMLObject in UMLDoc. This is NOT a long-term solution..I see
potential conflicts with badly implemented connections created in
umllistobject class for classifiers (why?) as well as possiblity that
you just *might* want to emit a childObjectCreated signal without then
adding that object to the UMLDoc (but seems unlikely). Also, the
add/removeStereoType methods got some warning messages... the
addStereoType method looks to be in conflict with set/getStereotype in
umlobject class and *should* be fixed soon.

svn path=/branches/KDE_3_2_BRANCH/kdesdk/umbrello/; revision=289733