backport fix from 27 Feb --
CVS commit by thomas:

Fix bug : 72615 : associations sometimes reverse because the
constuctor for the association widget was querying for any existing
associations that matched the one attempting to be created. IF one was
already present (such as when one tries to make more than one
aggregation between 2 classes) then it would try to swap the direction
of the association. Failing this, umlview would see that an
association already existed, and then prohibit adding the association
(correctly) but would then leave the object undeleted (incorrect). To
the extent that I have been able to test this patch, it works well,
however, more serious testing is needed to insure that this doesnt
generate new bugs elsewhere. It would be good if Oliver could review
this patch, and confer with me (bt) about what further steps are
needed.

svn path=/branches/KDE_3_2_BRANCH/kdesdk/umbrello/; revision=292245