patch by Achim Spangler:
the XMI load of message widgets is currently quite broken in CVS

The reason was a wrong logical combination in UMLView:
WAS:
-               if(!m_pDoc->loading() && !obj->activate(m_pDoc->getChangeLog()))
+{
SHOULD BE:
+               if(!m_pDoc->loading() || !obj->activate(m_pDoc->getChangeLog()))
+{

==>
The MessageWidget activation call was never performed during load of XMI.

==>
The lifeline of the objects weren't adopted to the message widget position AND
by the missing signal connection, each TRY to move a loaded message widget
caused a flip to the left margin of the QCanvas.

As the debug message of UMLWidget::adjustAssocs() was then activa again, I
used the KDebug function: KGlobal::kdBacktrace () from kdebug.h to get the
call stack for this unwanted call.

==>
The activate() method of MessageWidget called calculateDimensions(), which
then causes the call of UMLWidget::adjustAssocs(). As this calculation is
only needed for a new MessageWidget - and not during load - this call should
be made dependend on NotLoadingState.

With the attached diff, this call is performed only for new instances.

svn path=/trunk/kdesdk/umbrello/; revision=331508