patch by Achim Spangler:
part 1:
The load of EnumWidget caused unneeded adjustAssoc warnings. This was caused by the one and only constructor of EnumWidget which set always (!!) the size and triggered then the update, which then called adjust Assoc...

Now I simply inserted a check in the constructor, so that the size setting is only performed _without_ running load - i.e. on creation of really new instance.

part 2:
The other change is simply switching off ( comment out ) the warning line for each hide/show event in umlview.cpp

svn path=/trunk/kdesdk/umbrello/; revision=333503