BUG:57667 - A big round of applause to Ben Burton: <quote> The problem is essentially a misunderstanding of the drag/drop mechanism when working with trees.  The current umbrello code uses an incompatible mix between "drop onto the new parent" and "drop beneath the new parent" semantics.
The reason some people can't reproduce the bug is that, to reproduce it,
you need to drop the class over the upper half of the package item in
the logical tree (i.e., the mouse must end up over the upper half of the
text, not the lower half).  This causes umbrello to think that you want
to insert the class beneath the previous tree item intsead, since it
calculates "new parent node" using code that was really designed to
calculate "new final position of leaf" instead.  If you drop the class
over the lower half of the proper tree item, then umbrello behaves as
you would expect.

The fix is fortunately quite simple.  [...] Basically you ignore the
slotDropped() arguments altogether, since they correspond to movement
within a tree (drop over the final leaf position), not selection of a
tree node (drop over what will become the new parent node).

Thanks to Benjamin Mesing for providing a reproducible test case,
which made it possible to get on top of this one.
</quote>

svn path=/trunk/kdesdk/umbrello/; revision=398437