It seems that a previous version of Umbrello could create XMI files where some model objects were missing. This is a workaround so that Umbrello at least doesn't crash on loading such a file:
validateObjType(): Pass in the UMLObject* by reference. Add a Uml::IDType arg.
When the UMLObject* is NULL then create a new object with an artificial name,
"LOST_" followed by the id, and set its ID to the arg passed in.

This means that the user needs to manually rename the LOST_ items after loading
a corrupted file. Sorry.

BUG:145035

svn path=/branches/KDE/3.5/kdesdk/umbrello/; revision=661210