Check duplicates in packages only for objects which are not association types.
The code in UMLPackage::addObject() is a possible solution to Bug#142788 ( and BUG#138964 )
One refinement necessary is to enable cancelling the input of a new name, and thus cancel the operation.
But as the return value of this function is hardly checked anywhere in umbrello code, this would eventually lead to
duplicates again :(. Hence, one *has* to enter a new unique name and complete the operation.


svn path=/trunk/KDE/kdesdk/umbrello/; revision=696437