Fixed a bug which hapenned while moving widgets , which led to improper functionality of "Undo Move Widget"
The same variables m_oldX and m_oldY (which are supposed to only keep track of "old" static positions) were being 
used previously in UMLWidgetController to keep track of the previous X and Y positions in the dragged path ( the 
path being dynamic). New variables m_prevX and m_prevY are introduced for this purpose ( dynamic tracking) now.


svn path=/trunk/KDE/kdesdk/umbrello/; revision=696755