added import feature
If enabled in the new code import settings page for each imported c++ file
an artifact will be created in the component view.
The rational of this feature - when it is completed - is to generate files
completly from the parsed content including file header comments and the
exact collection of which classes belongs to a file.

To achieve this, a new class FileAST has been created, which holds all
file related informations not fitting into other declarations.

The implementation creates artifacts with the raw file name because there
are problems in Model_Utils::findUMLObject searching full pathes.

To complete this feature the following is to do:

- UMLArtifact has to be extended to hold uml objects collected for this file.

- ArtifactWidget has to be extended to let users editable the lists provided
  by UMLArtifact.

- The parser should add the collected uml objects to the related UMLArtifact

- The artifacts in the component view should be structured into folders using
  abolute pathes or based on the import root (switchable in the code import
  settings page)

svn path=/trunk/KDE/kdesdk/umbrello/; revision=1268827