Fixed bug variant of: Resizing synchronous message moves its position.
This commit fixes a variant of the related bug when loading message
widgets from a file. The shadow copy of widgets y position (m_nY)
has not been set.

The question is if this shadow copy of widget position is really required
or if it could be replaced by QGraphicsObject provided widget position.

CCBUG: 319813