Printing border fix.
Umbrello prints currently with screen resolution (96 dpi). This has the drawback that
QPrinter::pageRect() returns the same size as paperRect(), which in fact means that
there is no printing border.

To have a minimum border in that case we manually add a page border which is 2,5% of the
page width/height on each side.