Add visual feedback if widget has documentation.
This patch add a graphical indicator to widget painting, which is useful
to get a quick overview what is documentated and what not.

As graphical indicator a triangle in the lower left corner of the widget
is used, because it is still visible in lowest zoom level.

Widgets not derived from UMLWidget are not supported yet. MessageWidget
do not work because it do not have any documentation property.

GUI