Fix crash when adding entity constraints
The root cause of this crash is not checking for a valid index when retrieving
the item from the QList. We should at least check that value before calling at().

The patch I propose adds this check and also fixes the problem causing the
index to be wrong in the first place (like Arash Abedinzadeh described).

BUG: 191396
BUG: 237992
BUG: 257540
BUG: 302120
BUG: 322541