kdevcppparser: Always define the `identity' unary function.
_Identity is an internal, libstdc++-specific struct that is not to be relied
upon by external code.

preprocesslexer.cpp used to assume GCC (with libstdc++) and MSVC were the
only compilers available, and defined its own _Identity implementation for
the latter. It breaks things if any other compiler or STL implementation
(such as clang with libc++) is used.

The easiest fix is to just unconditionally define the `identity' unary
function and use it regardless of whether _Identity is defined or not.

Reviewed-by: Andi Fischer <andi.fischer@hispeed.ch>