Show only menu actions applicable for selection
When multiple widgets are selected, a context menu is shown containing only the
actions that work on the whole selection.

The old implementation had a few features like changing the "show" options
on a classifier for multiple selected widgets and bulk-change classifiers to
interfaces. That code is removed now since it wasn't used in the first place
and it's problematic in both implementation and user interface.

Additional improvements:
 * fix bug where fonts and colors sometimes wont apply
 * fix bug where right-click on an unselected widget does not select the widget
 * allow cloning multiple widgets
 * remove double keyboard shortcuts for delete action

Also, checkable menu items should not be used in the context menu of multiple
widgets. There's no sane initial value if the widgets are set to a different
value. The only nice way to handle this is have one menu item for setting the option
and one for unsetting. This is done for the "Use fill color" checkbox.

BUG: 329070