Fix various issues related to the command history
Summary of changes:
 * Track removal of umlobjects, widgets and diagrams in command history
 * Creating (and removing) diagrams can be undone
 * Commands always use IDs instead of pointers (potential fix for many related bugs)
 * Use macro's for moving and deleting multiple widgets
 * Clipboard actions use the command history (widget/umlobject creation)
 * Restore the CmdSetStereotype command
 * Cleanup useless change font/change multiplicity

UR commands reference diagrams and widgets by ID instead of keeping a pointer
to the objects. Pointers are unfit for the purpose of the commands because both
creation and removal of the objects can be undone - invalidating the pointers.

Because one UMLObject can have multiple widgets, the widgets can not be
identified by only the id(). To fix that issue, the localID from ObjectWidget
has been moved to UMLWidget. This does not change the existing behaviour of
ObjectWidgets but allows the UR commands to identify widgets that share the
same id.

REVIEW: 114744