Better fix for attribute creation crash
Previous fix was faulty. The problem occurs when an object is looked up with an
empty name. Instead of not finding an object, umldoc can return an arbitrary
association object instead, resulting in problems in Import_Utils.

This change prevents calling Import_Utils with an empty typeName, so when an
attribute is created without specifying a type, the default behaviour of
creating a new class with Object_Factory is triggered - which handles this case
well.

BUG: 225290