Additional history stack improvements
This commit tries to cleanup some actions in the history stack that do not
actually change anything or were not caused by a user action. The font metrics
on new widget fonts are applied before actually setting the font so we can
compare the old and new font to see if anything changed.

Also: fix datatype created with empty stereotype.