Add find text in tree view, current diagram and all diagrams feature.
The implementation adds menu entries and keyboard shortcuts for 'find', 'find next' and
'find previous' functions.

On start of a find operation a dialog will be opened, which allows to set the find filter
and a simple search text. Depending on the selected filter all tree view elements or elements
from the current or all diagrams are compared case insensitive against the search text.
The results are stored for later access and the first result will be displayed as described below.

Running 'find next' or 'find previous' using the 'tree view' filter will select the related search result,

Running 'find next' or 'find previous' with current diagram or all diagrams filter set, shows the related
view and selects the requested widget, which is positioned in the center of the view with a 100% zoom.
If a related tree view item is present, it will also be selected.

The implementation is undo/redo proof, because it stores scene/object/widget id's instead of pointers.

CCBUG:116354