Fix XMI 1.1 load/save issues
There were two problems causing the XMI file in the bug report to be
saved incorrectly, resulting in crashes after opening again. One problem
was association objects not being added to the document correctly. The other
problem was the nesting in the list view items: The classifiers were
nested in a folder that did not exist in the XMI. Fixed this by adding
the classifiers to the parent of the non-existent folder.

BUG: 334069
FIXED-IN: 4.13.1