Fix bug not placing class N1::A below N1 package in tree view.
The problem was, that Import_Utils::createUMLObject() tries to find an object
in different namespaces and only creates an uml object if not found.

In the case of parsing a c++ import class specification we simply need
to create the related uml class object in the current namespace, which
is provided by the additional parameter "doNotSearch".

Related test case is https://bugs.kde.org/show_bug.cgi?id=336810#c10

CCBUG:368810