Automate required namespace/package user interaction on c++ import.
If the type of an imported uml object could not be determined the import
creates an uml class object and flags the stereotype to be "class-or-package".
Such uml objects are displayed in the tree view by a dedicated icon.

If the real type of the related uml object is determined later as a class,
the flag is removed and the uml object is usable as a generic class.

If the real object is a namespace, the flag is removed and the uml object
is set to a package type and could be used as regular created package.
This is possible because the initial created object of type UMLClassifier
is derived from UMLPackage. In the latter case ClassifierWidget paints
the object as Package.

The tree view and classifier widget context menu has been extended to
provide functions to convert a "class or package" uml object to a class
or package.

Because the "class or package" state is flagged in the stereotype, saving
and loading of an umbrello project containing "class or package"
flagged uml objects is completly transparent.

CCBUG:336012