Fix creation of documentation for UML-Components
This fixes two issues connected to the DocWindow.

1. When DocWindow::showDocumentation is called, it checks whether it
is called with the same UMLObject/Scene/... But it missed that
UMLObjects and UMLScenes might be passed to the function embedded in a
widget.

2. The overwrite parameter of DocWindow::showDocumentation did the
opposite of what the documentation said. I changed it to match the
documentation and adjusted a few functions using it wrong.

BUG: 255424
BUG: 331215
FIXED-IN: 4.14.2
REVIEW: 120369