Don't change focus when documentation is shown.
When you select a widget in umlscene the documentation tab not just
gets shown, but also focused. This means that you are in editing mode
in this tab. So clicking on an element in the diagram and then pressing
DEL does not lead to deleting the element, what would be intuitive.
Instead you would delete text from the documentation tab if there is some.
This confusing behaviour occures with nearly all issues that have to do with
focusing. Leaving the focus on the selected element solves the problem.

REVIEW:125781