Add use of Object Diagram
Add code related to the object diagram actions
Fix part of the code wheres the warnings was with the
wrong association with the name of the class.

Add case related to Object Diagram, BUT i set the
Class Diagram Icon to this Object Diagram, still need to
decide what will be new icons for this diagram.

Fix Style issues

Signed-off-by: Lays Rodrigues <laysrodriguessilva@gmail.com>