Add all base for the instance/object diagram
Add the base for build the popup menu launched by the right
button, to access properties and make changes to the instance.
The use of wt_Object was changed to wt_Instace, to not conflit
with the wt_Object used in the Sequence Diagram.
The Instance Diagram is official called Object Diagram, and
I was using this like base, but was having some conflicts
with the base. Was better to change, so a lot of connections
around here related to the Object Diagram is called Instance.

Signed-off-by: Lays Rodrigues <laysrodriguessilva@gmail.com>