Class for Instance and Instance Attribute
Even with the attribute of the Object Diagram, being
a lot similar with the default attribute, I wasn't
able to manipulate the creation of the new attribute
in a good way. So following the idea of how is implemented
the Entity Attribute, I create this classes to handle that.
The instance.cpp to be the UMLInstance inherit from UMLClassifier
and the instanceattribute.cpp to be the UMLInstanceAttribute
inherit from UMLAttribute.
Overload the necessary methods.

Signed-off-by: Lays Rodrigues <laysrodriguessilva@gmail.com>