Base changes for use of instance's
The Diagram is called Object Diagram, but most of the
variables inside is named with *instance*. To not
conflict with the object used on Instance Diagram.
Since I wasnt able to load the Instance Attribute Dialog
based on the Attribute Dialog, I make it a class Instance
and Instance Attribute to handle this data.
All the changes here is base for it.

Signed-off-by: Lays Rodrigues <laysrodriguessilva@gmail.com>