New case tests for Object Diagram
Follow up the lvt_Attribute/ot_Attribute, searched around the code
places that need tests for the Object Diagram.
It is a shoot in the dark to attempt to fix the issues
with the Object Diagram. Until this point, you can create and
save to XMI, but when you load, the attributes of the Instance
isn't loaded.

Signed-off-by: Lays Rodrigues <laysrodriguessilva@gmail.com>