Fix creation and load of UMLInstance and UMLInstanceAttribute
Most of the errors related to the attributes of the Instances,
was because I used UMLAttribute instead UMLInstanceAttribute,
in the creation of new attributes.
Also, the load and save of attributes, is kind of confusing
to understand, but I fixed that, so now the Attributes
is saved and loaded with success.
Add in the UMLObject the loading of the instanceName,
if in the QDomElement has the keyword for it.

Signed-off-by: Lays Rodrigues <laysrodriguessilva@gmail.com>