Fix 'Font sizing is wrong when sharing diagrams'.
For any new file the current screen resolution is written into the xmi
file. If a file contains resolution the file coordinates are recalculated
into current screen resolution on load and recalculated to file resolution
on save.

Files without coordinates resolution are not recalculated on load or save,
because it is currently not known how to estimate the resolution the file
is created with. Adjusting those files requires currently adding the
"resolution" attribute to the <diagrams> tag manually.

To solve that issue an option would be to add a user request to get the
correct resolution or to print out a user warning on the status line after
loading and to add an option to be able to change the resolution in
document settings or if possible to estimate it from any diagram
coordinates.

BUG:90103
FIXED-IN:2.20.80 (KDE Applications 16.11.80)

Signed-off-by: Ralf Habacker <ralf.habacker@freenet.de>