Fix 'Lost class and interface names when open the project'.
UML objects used on diagrams may be located in the xmi file after the
diagram which results into "LOST_..." entries in the tree view. Caused by
internal limitations umbrello 2.18.x was not be able to fix that issues.
It was required to patch the xmi file by hand to remove the incorrect
listview tags.

To fix this issue there are three changes required:

1. Loading of diagrams is now performed after all uml objects has been
   loaded from the xmi file. On parsing diagram tags the related QDomNode
   and UMLFolder instance are saved in a map. After loading has been
   finished the diagrams are imported from this map.

2. Ignore uml objects from xmi file having no parent

3. Do not reparent uml objects caused by list view entries diverged
   from the uml model.

BUG:371495
FIXED-IN:2.20.3 (KDE Applications 16.08.03)


Signed-off-by: Ralf Habacker <ralf.habacker@freenet.de>