Fix 'Java import crashes on importing enums'.
This commit fixes the following issue: The importer creates UMLClassifier
objects on any unknown parsed type. On parsing the real definition later
it tries to use the UMLClassifier as UMLEnum which fails.

UMLDoc::setImporting() has been added to remove the old entry from the
tree view which is otherwise blocked in UMLListView::slotObjectRemoved()
by UmlDoc::loading().

BUG:373382
FIXED-IN:2.21.1 [KDE Applications 16.12.1)