Fix 'Crash in component diagram after deleting component'
In case a QGraphicsItem instance is on a scene and should be
deleted QGraphicsScenePrivate::removeItemHelper() is called.
The documentation say it leaves stale index in the tree, which let
QGraphicsScene crash on the next paint.

For now using the index has been disabled to avoid this issue.

FIXED-IN:2.23.1 (KDE Applications 17.08.1)
BUG:383592