Add macros to install translated artifacts for KF5
Those macros can be injected by release scripts, but we are moving
towards having them in the code directly. Moreover it would be
more complicated to inject them in the main CMakeLists.txt due to
the current hybrid nature of the umbrello code (it would require
an extra "if" for the KF5 case).
Please note that both functions are noop if the po/ directory does
not exist, so the default behavior is untouched.