Remove ifdef'ed translations to not block KF5 translation support
This is possible because umbrello/KF5 depends on KDELibs4Support.

In KAboutData instance setup calls to i18n() for translated
parameters are used ported back from KF5. For KDE4 builds those
calls are redefined to use ki18n().

I'm sorry to revert QCommandLineParser usage with this commit for
KF5 builds and to reintroduce deprecated API, but for my opinion
the additional required effort to maintain two incompatible
command line parser API's is not worth to stay with.

Initial I tried to write a simple wrapper class to hide command
line parser API differences between KDE4 and KF5 but I aborted
this approach because it would take too much work to fix all the
remaining type issues caused by the number of related classes.

Because the initial QCommandLineParser code will stay in git it
could be restored at any time in the future, so nothing is lost.