Fix 'Namespace mismatch of types imported from java import between object and tree view'
Unknown types imported from java files are assigned initial to the global namespace.
If the related type is found later, the namespace will be updated. This update was not
propagated to the tree view.

To propagate the update to the tree view it is required to enable the UMLObject::modified()
signal temporary on changing the package parent.

The signal emitting has been moved down in code to be sure that all other changes are applied.

BUG:386698
FIXED-IN:2.23.80 (KDE Applications 17.11.80)