Add manual unit test for class ListPopupMenu
To be able to see the results of related code change a unit test has been
added, which dumps out all possible context menus.

For specific tests the debug support for class ListPopupMenu through the
debug window has been extended to dump the context of the currently opened
menu to the configured debug output.