Fixes broken docbook URLs when creating xhtml
The stylesheet docbook.xsl no longer exists under ksgmtools2,
but is included in the docbook package and the remote docbook.xsl
url no longer fits.

The code for determining user-defined and local docbook xsl files
has been moved to the XhtmlGenerator class in addition to the
existing queries.

BUG:400657
FIXED-IN:2.26.3 (KDE Applications 18.08.3)