file import: do not use overwritten temporary file
We checked if the temporary file we created still exists after its creation,
yet in the meantime we have already copied to it (which overwrites it):
KIO::FileCopyJob *job = KIO::file_copy(url, dest, -1, KIO::Overwrite);

Instead of the temporary file, check if the regular file we use later on
exists.

This bug is a result of fixing bug 125102.

BUG:406574
FIXED-IN:2.28.0 (KDE Applications 19.04.0)
CCBUG:125102

Reviewed By: habacker
Differential Revision: https://phabricator.kde.org/D20775