Fix 'No context menu entry for adding messages to sequence diagram'
With this patch UMLscene::m_pToolBarState and m_pToolBarStateFactory has
been moved into class UMLScenePrivate for easier access.

The "found" message is added to the context menu if no object line is
selected, all other message requires to open the context menu on the
widget line of an object.

BUG:409969
FIXED-IN:2.28.80 (KDE Applications 19.07.80)