Add UML2 support for "required interface" associations
With UML2 a "required interface" association is created by adding a
normal association from the interface to the component, wherein adding
an association from the component to the interface creates a "provided
interface".

With UML1, according to current research, there is no visible label for
a required interface mapping. This behavior has been corrected.

CCBUG:115269