Fix 'Classifier unit case not enabled'
One problem was that comparing the list address of contained objects
in UMLCanvasObject is incorrect. Instead, each object in this list
must be compared. Also, the comparison of UMLClassifier pointers
does not seem to give the desired result.

Unimplemented tests are now skipped to avoid aborting the test.

BUG:414044
FIXED-IN:2.29.90 (KDE Applications 19.11.90)