Move WidgetBase::addAssoc(), removeAssoc() and m_nlocalID to class WidgetBase
To be able to use a widget that is linked to another diagram as a link
end in the AssociationWidget class, some methods must be encapsulated.
In order to use the class WidgetBase as base class of such a
"proxy widget", some methods and class members located in the UMLWidget
have to be moved to WidgetBase.

CCBUG:53369