Improve auto layout for associations
The line type selected in the association settings dialog is passed to
the 'dot' command line tool, which creates the lines according to the
selected style. The modified lines then get this selected type and the
imported (control) points are passed to the lines.

When importing polylines, duplicate points are removed.

BUG:414595
FIXED-IN:2.30.80 (KDE Applications 20.03.80)