Add support for sequence diagrams to destroy an object
For the new message type a corresponding icon has been added to the
toolbar and a corresponding entry to the context menu.

The version of the xmi file was increased because a new value was
added to the xml tag for message types.

The api of the class SeqLineWidget for setting the line color was
unified by adding a method setLineColorCmd().

BUG:418249
FIXED-IN:2.30.80 (KDE releases 20.03.80)