Fixed missing entries for showing/hiding in the settings menu
The corresponding menu entries were not displayed in the settings
menu because the associated actions were initialized too late.
They must be present before the umbrelloui.rc file is loaded.

To fix the problem, the initialization of actions in the
UMLAppPrivate class has been moved to a new method
initActions() and the window creating moved from
UMLApp::initView() to UMLAppPrivate::initWidgets().

Furthermore setting the visual state of the related menu
entries has also been fixed.

CCBUG:508513