Fix bug not deleting uml association
When an association widget is deleted, the associated UML
association is not deleted. The remaining associations
are no longer visible because they do not appear in the
property dialog of a class. Since they are no longer
visible, cyclic inheritances can easily occur, which
can lead to a crash, e.g., of the Java export.

After this fix, it is important to restore the missing
widgets for such associations and to point out cyclic
inheritances when loading files.

BUG:511328
FIXED-IN:25.11.80